Updated topic-selector to new UI format, still need to add confirmation messages when deleting items from database

+++ clearSelections.php
@@ -1,30 +1,45 @@
 <?php
 require_once "../config.php";
 
-use Tsugi\Core\LTIX;
+use \Tsugi\Core\LTIX;
 
 $p = $CFG->dbprefix;
 
 $LAUNCH = LTIX::requireData();
 
+include("menu.php");
+
+$OUTPUT->header();
+
 $OUTPUT->bodyStart();
 
-$topicST  = $PDOX->prepare("SELECT * FROM {$p}topic");
-$topicST->execute(array());
-$topic = $topicST->fetchAll(PDO::FETCH_ASSOC);
+$OUTPUT->topNav($menu);
 
-$numReserved = 0;
+$OUTPUT->flashMessages();
 
-foreach($topic as $top) {
-    $newTopic = $PDOX->prepare("UPDATE {$p}topic SET num_reserved=:numReserved WHERE topic_id = :topicId");
-    $newTopic->execute(array(
-        ":topicId" => $top['topic_id'],
-        ":numReserved" => $numReserved,
-    ));
+if(isset($_GET['confirm'])) {
+    if(($_GET['confirm'] == 2)) {
+        $topicST = $PDOX->prepare("SELECT * FROM {$p}topic");
+        $topicST->execute(array());
+        $topic = $topicST->fetchAll(PDO::FETCH_ASSOC);
+
+        $numReserved = 0;
+
+        foreach ($topic as $top) {
+            $newTopic = $PDOX->prepare("UPDATE {$p}topic SET num_reserved=:numReserved WHERE topic_id = :topicId");
+            $newTopic->execute(array(
+                ":topicId" => $top['topic_id'],
+                ":numReserved" => $numReserved,
+            ));
+        }
+
+        $delSelections = $PDOX->prepare("DELETE FROM {$p}selection");
+        $delSelections->execute(array());
+
+        header('Location: ' . addSession('index.php'));
+        return;
+    }
 }
 
-$delSelections = $PDOX->prepare("DELETE FROM {$p}selection");
-$delSelections->execute(array());
-
-header('Location: ' . addSession('index.php'));
+header('Location: ' . addSession('index.php?confirm=2'));
 return;
